clean up Rails/FilePath, Rails/OutputSafety and Style/GuardClause
