+++ lib/bootstrap_form/helpers/bootstrap.rb
@@ -24,35 +24,34 @@ module BootstrapForm
 
       def alert_message(title, options={})
         css = options[:class] || "alert alert-danger"
+        return unless object.respond_to?(:errors) && object.errors.full_messages.any?
 
-        if object.respond_to?(:errors) && object.errors.full_messages.any?
-          content_tag :div, class: css do
-            concat content_tag :p, title
-            concat error_summary unless options[:error_summary] == false
-          end
+        content_tag :div, class: css do
+          concat content_tag :p, title
+          concat error_summary unless options[:error_summary] == false
         end
       end
 
       def error_summary
-        if object.errors.any?
-          content_tag :ul, class: "rails-bootstrap-forms-error-summary" do
-            object.errors.full_messages.each do |error|
-              concat content_tag(:li, error)
-            end
+        return unless object.errors.any?
+
+        content_tag :ul, class: "rails-bootstrap-forms-error-summary" do
+          object.errors.full_messages.each do |error|
+            concat content_tag(:li, error)
           end
         end
       end
 
       def errors_on(name, options={})
-        if has_error?(name)
-          hide_attribute_name = options[:hide_attribute_name] || false
+        return unless has_error?(name)
 
-          content_tag :div, class: "alert alert-danger" do
-            if hide_attribute_name
-              object.errors[name].join(", ")
-            else
-              object.errors.full_messages_for(name).join(", ")
-            end
+        hide_attribute_name = options[:hide_attribute_name] || false
+
+        content_tag :div, class: "alert alert-danger" do
+          if hide_attribute_name
+            object.errors[name].join(", ")
+          else
+            object.errors.full_messages_for(name).join(", ")
           end
         end
       end
@@ -81,12 +80,12 @@ module BootstrapForm
       def prepend_and_append_input(name, options, &block)
         options = options.extract!(:prepend, :append, :input_group_class)
 
-        input = capture(&block) || "".html_safe
+        input = ActiveSupport::SafeBuffer.new(capture(&block) || "")
 
         input = prepend_input(options) + input + append_input(options)
         input += generate_error(name)
         options.present? &&
-          input = content_tag(:div, input.html_safe, class: ["input-group", options[:input_group_class]].compact)
+          input = content_tag(:div, input, class: ["input-group", options[:input_group_class]].compact)
         input
       end
 
@@ -109,21 +108,21 @@ module BootstrapForm
 
       def append_input(options)
         html = content_tag(:div, input_group_content(options[:append]), class: "input-group-append") if options[:append]
-        (html || "").html_safe
+        ActiveSupport::SafeBuffer.new(html || "")
       end
 
       def prepend_input(options)
         html = content_tag(:div, input_group_content(options[:prepend]), class: "input-group-prepend") if options[:prepend]
-        (html || "").html_safe
+        ActiveSupport::SafeBuffer.new(html || "")
       end
 
       def setup_css_class(the_class, options={})
-        unless options.key? :class
-          if (extra_class = options.delete(:extra_class))
-            the_class = "#{the_class} #{extra_class}"
-          end
-          options[:class] = the_class
+        return if options.key? :class
+
+        if (extra_class = options.delete(:extra_class))
+          the_class = "#{the_class} #{extra_class}"
         end
+        options[:class] = the_class
       end
     end
   end
