+++ lib/bootstrap_form/inputs.rb
@@ -5,6 +5,7 @@ module BootstrapForm
     extend ActiveSupport::Autoload
 
     autoload :Base
+    autoload :InputsCollection
     autoload :CheckBox
     autoload :CollectionCheckBoxes
     autoload :CollectionRadioButtons
